requirements.txt
Django>=4.2
firebase-admin>=6.0.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
Pillow>=10.0.0
requests>=2.32.0
whitenoise>=6.7.0
